requirements.txt
aiogram==2.25.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
Pillow==10.0.1
pytesseract==0.3.10
openai==1.3.0
rarfile==4.0 